Added real time notifications and improved querying

+++ app/Http/Controllers/UserController.php
@@ -2,43 +2,66 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\NewEmailRequest;
 use App\Http\Requests\UpdateProfileRequest;
+use App\Models\Email;
 use App\Models\User;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\File;
 use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
-	public function userData()
+	public function index(): JsonResponse
 	{
-		return response()->json(auth()->user(), 200);
+		$user = jwtUser();
+		if ($user)
+		{
+			$userData = [
+				'id'       => $user->id,
+				'name'     => $user->name,
+				'email'    => $user->email,
+				'thumbnail'=> $user->thumbnail,
+				'google_id'=> $user->google_id,
+				'emails'   => $user->emails,
+			];
+		}
+		else
+		{
+			$userData = null;
+		}
+		return response()->json(['user' => $userData]);
 	}
 
-	public function update(UpdateProfileRequest $request)
+	public function update(UpdateProfileRequest $request): JsonResponse
 	{
 		$data = $request->validated();
-		$email = auth()->user()->email;
-		$currentThumbnail = auth()->user()->thumbnail;
-		$thumbnail = $request->file('thumbnail')->store('images');
+		$email = jwtUser()->email;
+		$currentThumbnail = jwtUser()->thumbnail;
+		if($request->file('thumbnail')){
+			$thumbnail = $request->file('thumbnail')->store('images');
+		}else{
+			$thumbnail=$currentThumbnail;
+		}
 
 		if ($request->password)
 		{
 			$data['password'] = bcrypt($data['password']);
 		}
 
-		$code = auth()->user()->verification_code;
+		$code = jwtUser()->verification_code;
 
-		if ($request->email != $email && $request->email)
+		if (($request->email != $email) && $request->email)
 		{
-			$url = env('FRONTEND_URL') . '/update-email/' . $code . '?email=' . $request->email;
+			$url = env('FRONTEND_URL_FOR_CONFIRM') . '/update-email/' . $code . '?email=' . $request->email;
 			$body = 'You asked for Email change? then change it.';
 			$buttonText = 'Change email';
 			EmailVerificationController::sendVerifyEmail($request->email, $code, 'Change Email', 'emails.reset', $body, $buttonText, $url);
 			$data['is_verified'] = 0;
 		}
 		$data['email'] = $email;
-		if ($currentThumbnail && $currentThumbnail != 'assets/LaracastImage.png' && $thumbnail)
+		if ($currentThumbnail && $currentThumbnail != 'assets/LaracastImage.png' && $thumbnail && $thumbnail!=$currentThumbnail)
 		{
 			$absolutePath = storage_path('/app/' . $currentThumbnail);
 			File::delete($absolutePath);
@@ -53,15 +76,32 @@ class UserController extends Controller
 			$data['thumbnail'] = $currentThumbnail;
 		}
 		User::where('email', $email)->update($data);
-		return response()->json($data, 200);
+		return response()->json($data);
 	}
 
-	public function submitChangeEmail(Request $request)
+	public function submitChangeEmail(Request $request): JsonResponse
 	{
 		if ($request->email)
 		{
 			DB::table('users')->where('verification_code', $request->token)->update(['email'=>$request->email, 'is_verified'=>1]);
 		}
-		return response()->json('Email changed successfully!', 200);
+		return response()->json('Email changed successfully!');
+	}
+
+	public function addNewEmail(NewEmailRequest $request): JsonResponse
+	{
+		Email::create([
+			'user_id'=> jwtUser()->id,
+			'address'=> $request->new_email,
+		]);
+		$user = jwtUser();
+		EmailVerificationController::sendEmail($user->name, $request->new_email, $request->new_email, 'Account Confirmation', 'emails.register');
+		return response()->json(Email::where('user_id', jwtUser()->id)->get());
+	}
+
+	public function destroy(Email $email): JsonResponse
+	{
+		$email->delete();
+		return response()->json(Email::where('user_id', jwtUser()->id)->get());
 	}
 }
